Extract NSR into dedicated class
